Fix problem where field trials using kExpirationYearInFuture could get disabled.
The issue was that kExpirationYearInFuture was using system time, but the
expiration check in field_trial.cc was using build time. This CL changes
the constant to be named kExpirationYearNotExpired and sets it based on
build time.

BUG=168799
TEST=New unit test.

Review URL: https://chromiumcodereview.appspot.com/11783033

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d0c692981d34a0aa35aea070eabcae656ee5ef74