cc: Set ui/gpu/cc/ipc/upload/raster thread priorities.
This bumps up the ui/gpu/cc/ipc threads and bumps down the
upload/raster threads. The webkit thread is left alone, as
it needs to by dynamically adjusted (in another patch).

To reduce risk and for easier merging we do this as #ifdefs
and only for Android. This bug tracks abstracting the code
and doing it on all platforms: http://crbug.com/170549

BUG=164651

Review URL: https://chromiumcodereview.appspot.com/11880014

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9875de7440186a564a15f2399be0c32434b9c40