Android: fix race condition in jni_generator.
Instead of checking if the directory exists and then creating it, just
try and create it and catch the exception if it already exists. This
fixes a race that seems to occur when using the Android build system for
WebView, as it appears to run several instances of jni_generator at
exactly the same time.

BUG=

Review URL: https://codereview.chromium.org/12461009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 126b7060ab84793646b33307a842895fcdcf9cf9