JNI on Android: refresh self-documenting "SampleForTests.java".
This file is the authoritative source of documentation and information on how to
generate JNI bindings.
However, it was probably too specific and probably required a fair bit of previous knowledge.
This is an attempt in making it more comprehensive, aiming to have these files as a full guide
for all supported cases and the best practices, and help everyone work with JNI more effectively.

BUG=
TEST=jni_generator_tests


Review URL: https://chromiumcodereview.appspot.com/12600014

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b739734e9d1ec8a32a6fce3711a3439329b9eb5a