When running iOS tests on devices, write stdout to a file, then dump to NSLog.
gtest only writes test output to stdout, but on iOS 6 devices, stdout is not
available to be retrieved from the device. This CL redirects stdout to a file
while the tests are running, then dumps the contents of that file to NSLog
(which is available to be retrieved from the device).

BUG=None


Review URL: https://chromiumcodereview.appspot.com/12328083

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ea346d9d6b25ad37eeafbc9840ab4626c8366f9