Forward-declare enums in a way that is compatible with NS_ENUM
Starting with the 10.8 SDK and the iOS 6.0 SDK, enums are declared
strongly typed in C++11 mode, which the old forward-declares weren't
compatible with. This creates a new macro for forward-declaring these
enums in a way that will match the SDK and compilation mode.

BUG=226494

Review URL: https://chromiumcodereview.appspot.com/13455004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 965bcb00895e398c725d79915ecb0e92321e8a66