Android: improve native debuggability
When starting with --wait-for-debugger command line option, it's ending
up in a busy loop waiting for a variable to be changed. Change that busy
loop from a macro to a function. It avoids having to step on instruction
level to find the right stack frame before the variable is visible.


Review URL: https://chromiumcodereview.appspot.com/13597005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 901a5e3835718e8872f38af32d65fce7ea3fcbeb