Update ProcessIterator::CheckForNextProcess to get exe name from /proc/pid/exe.
The startup performance_ui_tests have been hanging on the perf bots since they
were upgraded to precise. I tracked this down to GetRunningChromeProcesses()
not returning all of the chrome processes so the ProxyLauncher can't kill
them all. When the ProxyLauncher doesn't kill all child chrome processes, the
buildbot step hangs.

It turns out that when the test hangs, there is a chrome process which has
correct entries in /proc/pid/cmdline and /proc/pid/exe, but the short executable
name in /proc/pid/stat is "dconf worker" instead of "chrome". This fools the
NamedProcessIterator into not matching it as a chrome process.

This incorrect information in /proc/pid/stat is bizarre and I have no idea what
the root cause is. But since there was a TODO here anyway to get the process
name from cmdline instead of stat, I went ahead and did that as a workaround.

BUG=235893
TEST=Startup performance_ui_tests and base_unittests
R=jar@chromium.org, thestig@chromium.org

Review URL: https://codereview.chromium.org/14561002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fd940fcb22ba591ed1f24b91de77457c8558b14b