Add more error codes to base::CreatePlatformFile on windows.
This is a rough classification of the errors that showed up in the
PlatformFile.UnknownCreateFileErrors histogram. I added an IO entry to
to the enum because a bunch of the errors seem to indicate a low-level
problem that is out of the application programmer's control.

BUG=229252
R=jar@chromium.org

Review URL: https://codereview.chromium.org/14586003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 065499f70249cd46f7c04e727e90207179e39408