Allow tracing categories to default to disabled.
This CL allows the prefixing of category names with disabled- to have them
not traced by default. Only if an explicit disabled-foo is provided in the
category filter do we enable the category.

Providing a disabled-foo in the category filter does not act like an include
category in that all other includes have to be specified. The disables are
treated as their own group.

BUG=232109

Review URL: https://chromiumcodereview.appspot.com/14748003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aaffef4e38d4fc42a6424630182d500cba32eaca