moving SyncSocket's kInvalidHandle constant
So it does not get exported in the split dll build.

The change is to move the definition into the h and
telling the linker to just pick one symbol.

This change needs to be undone in the next quarter
or so.

BUG=237249
TEST=none
TBR=brettw

Review URL: https://chromiumcodereview.appspot.com/15480004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0f248ef374b7a02daf9227cdc6d7d852159183fd