Move latin1_string_conversions to base
Unfortunately, we can't keep this function in webkit/glue because not everyone
who links in Blink depends on webkit/glue. For example, cc uses WebString but
doesn't depend on webkit/glue.

IMHO, the best solution to all these constraints is to put this function in
base. It's concerned with concepts that make sense in base (strings and
character sets). The only thing odd about the function is that it accepts
Latin-1 or UTF-16, which is because of the underlying string representation in
Blink.

Review URL: https://chromiumcodereview.appspot.com/15732022

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45dba2731166c1dfd7093b08afa22f73afc5f3c3