Adds RefCountedDeleteOnMessageLoop
RefCountedDeleteOnMessageLoop is similar to RefCountedThreadSafe, but guarantees that
the referenced object will be deleted on the specified loop.
The functionality is similar to content::DeleteOnThread<> but does not depend on
BrowserThread.

The CL includes an example use of the class.

BUG=248049
TBR=stuartmorgan, joth

Review URL: https://chromiumcodereview.appspot.com/16938006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f0c46213502eec4aa27aa9372323c66e2ee0224