Do not show sync promo when RestoreOnStartupURLs policy is set
When first_run_tabs are set through master_preferences, the sync promo
should be shown if not explicitly suppressed. When RestoreOnStartupURLs
is set through policy, the sync promo should never be shown. This is
long-standing behavior, as evidenced here:

http://crbug.com/125467#c26

That behavior recently regressed with the sync promo appearing even if
RestoreOnStartupURLs are set. The regression happened here:

https://chromiumcodereview.appspot.com/12638005

This CL fixes the regression by ensuring that when RestoreOnStartup=4 and
RestoreOnStartupURLs are set, the sync promo is not shown. The CL also
adds browser tests to protect against similar regressions in the future.

BUG=244849
TEST=Manual and new browser_tests

Review URL: https://chromiumcodereview.appspot.com/16141008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93660ab3f5c895eb3e8a5f66d3af4669e16de21d