content_shell: Use a ThreadRestrictions::ScopedAllowWait for layout tests on Android.
Because of the nested message pump implementation used by Android for
running layout tests, which uses a base::WaitableEvent to wait for new
messages, we need to allow waiting on the UI thread.

The ScopedAllowWaitForAndroidLayoutTest class has been implemented as
a member of the WebKitTestController, to ensure that the exception's
lifetime is tied to the layout test controller itself.

BUG=232044
R=jam@chromium.org, jochen@chromium.org

Review URL: https://codereview.chromium.org/17395002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 24ad9ea17ebf454c7d39ba5f429b0f679a4f8a84