Add ScopedPlatformFileCloser to base.
There are several places that need to be sure to close a platform file and have implemented a scoped to ensure that it is done.  Instead of adding yet another version of this code, this moves it into base and updates all known implementations to use it.

BUG=NONE

Review URL: https://chromiumcodereview.appspot.com/20154002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a4b577d5b84e09dec6937c34c9400cdeed59eaf