gtk: Allow building both the X11 and Gtk message-pumps for gtk.
This patch allows both the X11 and the Gtk message-pumps to be built
in a linux-gtk build. A subsequent patch will use the X11 message-pump
for the GPU process, while continue to use the Gtk message-pump for the
browser process.

BUG=145600
R=ccameron@chromium.org, oshima@chromium.org, piman@chromium.org, thakis@chromium.org

Review URL: https://codereview.chromium.org/23880006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb99d70f77e537ffc7ec0921d90c13e3cf689615