Copy the thread name to the stack for debugging purposes.
The this ptr in Thread is getting smashed after the MessageLoop ends, which
leaves very little to go on. This is kind of a stab in the dark, but hopefully
it makes something bleed in a general direction.

BUG=274705
TEST=Minidumps contain 47 bytes of thread name data way high on the stack.
R=mark@chromium.org

Review URL: https://codereview.chromium.org/23606030

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58f036475e1f6fbfb9c02d9c86176aa49610584a