base: Add base::TimeTicks::IsHighResNowFastAndReliable
This will allow clients determine if it's okay to call
HighResNow often and rely on its results. If it is not
reliable, the client can use a fallback path that is
better than relying on a bad HighResNow.

BUG=305645

Review URL: https://codereview.chromium.org/27317002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cef72a9c2505d9de8db889a43c8cf8f9f44baf4b