base: Remove call to g_get_home_dir() from GetHomeDir() for Chrome OS
This is in preparation for removing a dependency to GLib.

On a Chrome device, home directory is always set to /home/chronos/user
as hard-coded in GetHomeDir().
On a linux workstation for testing, referencing $HOME is sufficient.

BUG=240617

Review URL: https://codereview.chromium.org/60083009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4ab65012b06cd2cfdb832819f3b2a61369db773f