Clean up uses of use_glib that are not related to glib dependencies
We're frequently using the use_glib gyp variable as a shorthand for:

  OS!="win" and OS!="mac" and OS!="ios" and OS!="android"

This should be equivalent to:

  desktop_linux == 1 or chromeos == 1

Embedded builds want to remove as many dependencies as possible. This
cleans up uses of the use_glib variable so we can turn it off on linux
without breaking the build.

BUG=316226

Review URL: https://codereview.chromium.org/61923004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b098e1d8b492803cbb2857f684fdc34460fdc9e1