Fix threading in ScopedCriticalAction
The locking code doesn't actually work as written, because if the
expiration has already fired, but not started running EndBackgroundTask,
when the destructor runs, then both the lock and the task ID are
pointers into a deallocated object, and potentially garbage.

To fix it, move the logic to a ref-counted inner object, so that the
expiration handler will keep the object alive.

BUG=321667

Review URL: https://codereview.chromium.org/81513003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68a5643d4b33a7a284a2f1dd408d83a79ff6672f